remove xrun feature check
I forgot openASIO wrapper is missing this value.
xruns is not used significantly in Juce so far as I can tell/remember, also
the feature should still be functional if available because the message
handler enum is there. The trick is "if" a device can't report the number
the value remains at 0 (when in code it might need to be set to -1)